+++ plugins/emoji/plugin.js
@@ -379,10 +379,11 @@
 					return arrTools.reduce(
 						items,
 						function( acc, item ) {
+							addEncodedName( item );
 							return acc + emojiTpl.output( {
 									symbol: htmlEncode( item.symbol ),
 									id: htmlEncode( item.id ),
-									name: htmlEncode( item.id.replace( /::.*$/, ':' ).replace( /^:|:$/g, '' ).replace( /_/g, ' ' ) ),
+									name: item.name,
 									group: htmlEncode( item.group ),
 									keywords: htmlEncode( ( item.keywords || [] ).join( ',' ) )
 								} );
@@ -599,7 +600,7 @@
 					editor._.emoji.autocomplete = new CKEDITOR.plugins.autocomplete( editor, {
 						textTestCallback: getTextTestCallback(),
 						dataCallback: dataCallback,
-						itemTemplate: '<li data-id="{id}" class="cke_emoji-suggestion_item">{symbol} {id}</li>',
+						itemTemplate: '<li data-id="{id}" class="cke_emoji-suggestion_item"><span>{symbol}</span> {name}</li>',
 						outputTemplate: '{symbol}'
 					} );
 				}
@@ -641,6 +642,7 @@
 								return a.id > b.id ? 1 : -1;
 							}
 						} );
+					data = arrTools.map( data, addEncodedName );
 					callback( data );
 				}
 			} );
@@ -657,6 +659,13 @@
 
 		}
 	} );
+
+	function addEncodedName( item ) {
+		if ( !item.name ) {
+			item.name = htmlEncode( item.id.replace( /::.*$/, ':' ).replace( /^:|:$/g, '' ) );
+		}
+		return item;
+	}
 } )();
 
 /**